pyproject.toml
[project]
name = "po3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clip",
    "easyocr>=1.7.2",
    "einops>=0.8.2",
    "elevenlabs>=2.40.0",
    "fastapi>=0.135.1",
    "faster-whisper>=1.2.1",
    "google-cloud-vision>=3.12.1",
    "google-genai>=1.68.0",
    "jupyter>=1.1.1",
    "librosa>=0.11.0",
    "mediapipe==0.10.14",
    "numpy>=2.4.3",
    "ollama>=0.6.1",
    "openai>=2.29.0",
    "openai-whisper>=20250625",
    "opencv-contrib-python>=4.13.0.92",
    "pillow>=12.1.1",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "python-multipart>=0.0.22",
    "qwen-tts>=0.1.1",
    "roboflow>=1.2.16",
    "scikit-learn>=1.8.0",
    "soundfile>=0.13.1",
    "speechrecognition>=3.15.1",
    "streamlit>=1.55.0",
    "timm>=1.0.26",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=4.47.0",
    "translate>=3.8.0",
    "ultralytics>=8.4.22",
    "utils>=1.0.2",
    "uvicorn>=0.42.0",
    "xgboost>=3.2.0",
]

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
clip = { git = "https://github.com/openai/CLIP.git" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
